+++ tests/cpp-tests/Classes/Camera3DTest/Camera3DTest.cpp
@@ -900,7 +900,7 @@ void CameraCullingDemo::addSpriteCallback(Ref* sender)
     
     // update sprite number
     char szText[16];
-    sprintf(szText,"%ld sprits", static_cast<long>(_layer3D->getChildrenCount()));
+    sprintf(szText,"%ld sprits", static_cast<int32_t>(_layer3D->getChildrenCount()));
     _labelSprite3DCount->setString(szText);
 }
 
@@ -928,7 +928,7 @@ void CameraCullingDemo::delSpriteCallback(Ref* sender)
     
     // update sprite number
     char szText[16];
-    sprintf(szText,"%ld sprits", static_cast<long>(_layer3D->getChildrenCount()));
+    sprintf(szText,"%ld sprits", static_cast<int32_t>(_layer3D->getChildrenCount()));
     _labelSprite3DCount->setString(szText);
 }
